+++ node/src/test/test-werift-sctp.ts
@@ -221,20 +221,13 @@ test('ordered DataProducer delivers all SCTP messages to the DataConsumer', asyn
 		},
 	]);
 
-	// console.log(
-	// 	'--- getBufferedAmount():',
-	// 	await ctx.dataConsumer!.getBufferedAmount()
-	// );
-
-	for (let i = 0; i < 100000; i++) {
-		const bufferedAmount = await ctx.dataConsumer!.send('hola');
-		if (bufferedAmount > 0) {
-			console.log('--- bufferedAmount:', bufferedAmount);
-		}
-	}
-
-	console.log(
-		'--- FIN getBufferedAmount():',
-		await ctx.dataConsumer!.getBufferedAmount()
-	);
+	expect(await ctx.dataConsumer!.getBufferedAmount()).toBe(0);
 }, 10000);
+
+test('can send messages from the SCTP DataConsumer', async () => {
+	expect(await ctx.dataConsumer!.getBufferedAmount()).toBe(0);
+
+	// We can send messages directly from the DataConsumer and it should return
+	// current buffered amount, so 0 bytes.
+	await expect(ctx.dataConsumer!.send('foo')).resolves.toBe(0);
+}, 2000);